+++ R/other_useful_functions.R
@@ -456,7 +456,28 @@ convertBlockList <- function(fileNameBlocksDetPlink,
 #' then the `addNOIA` (or corresponding) option in the `calcGRM` function will be used,
 #' and if this argument is `phylo`, the gaussian kernel based on phylogenetic distance will be computed from phylogenetic tree.
 #' You can assign more than one kernelTypes for this argument; for example, kernelTypes = c("addNOIA", "phylo").
-#' @param nCores The number of cores used for optimization.
+#' @param n.core Setting n.core > 1 will enable parallel execution on a machine with multiple cores.
+#' This argument is not valid when `parallel.method = "furrr"`.
+#' @param parallel.method Method for parallel computation. We offer three methods, "mclapply", "furrr", and "foreach".
+#'
+#' When `parallel.method = "mclapply"`, we utilize \code{\link[pbmcapply]{pbmclapply}} function in the `pbmcapply` package
+#' with `count = TRUE` and \code{\link[parallel]{mclapply}} function in the `parallel` package with `count = FALSE`.
+#'
+#' When `parallel.method = "furrr"`, we utilize \code{\link[furrr]{future_map}} function in the `furrr` package.
+#' With `count = TRUE`, we also utilize \code{\link[progressr]{progressor}} function in the `progressr` package to show the progress bar,
+#' so please install the `progressr` package from github (\url{https://github.com/HenrikBengtsson/progressr}).
+#' For `parallel.method = "furrr"`, you can perform multi-thread parallelization by
+#' sharing memories, which results in saving your memory, but quite slower compared to `parallel.method = "mclapply"`.
+#'
+#' When `parallel.method = "foreach"`, we utilize \code{\link[foreach]{foreach}} function in the `foreach` package
+#' with the utilization of \code{\link[parallel]{makeCluster}} function in `parallel` package,
+#' and \code{\link[doParallel]{registerDoParallel}} function in `doParallel` package.
+#' With `count = TRUE`, we also utilize \code{\link[utils]{setTxtProgressBar}} and
+#' \code{\link[utils]{txtProgressBar}} functions in the `utils` package to show the progress bar.
+#'
+#' We recommend that you use the option `parallel.method = "mclapply"`, but for Windows users,
+#' this parallelization method is not supported. So, if you are Windows user,
+#' we recommend that you use the option `parallel.method = "foreach"`.
 #' @param hOpt Optimized hyper parameter for constructing kernel when estimating haplotype effects.
 #'  If hOpt = "optimized", hyper parameter will be optimized in the function.
 #'  If hOpt = "tuned", hyper parameter will be replaced by the median of off-diagonal of distance matrix.
@@ -528,7 +549,8 @@ estPhylo <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.set
                      distMat = NULL, distMethod = "manhattan", evolutionDist = FALSE,
                      subpopInfo = NULL, groupingMethod = "kmedoids",
                      nGrp = 3, nIterClustering = 100, kernelTypes = "addNOIA",
-                     nCores = parallel::detectCores() - 1, hOpt = "optimized",
+                     n.core = parallel::detectCores() - 1,
+                     parallel.method = "mclapply", hOpt = "optimized",
                      hOpt2 = "optimized", maxIter = 20, rangeHStart = 10 ^ c(-1:1),
                      saveName = NULL, saveStyle = "png",
                      pchBase = c(1, 16), colNodeBase = c(2, 4), colTipBase = c(3, 5, 6),
@@ -785,23 +807,16 @@ estPhylo <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.set
 
 
             if (length(hStarts) >= 2) {
-              if (verbose) {
-                solnList <- pbmcapply::pbmclapply(X = hStarts,
-                                                  FUN = function(h) {
-                                                    soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
-                                                                   lower = 0, upper = 1e06, control = list(iter.max = maxIter))
+              solnList <- parallel.compute(vec = hStarts,
+                                            func = function(h) {
+                                              soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
+                                                             lower = 0, upper = 1e06, control = list(iter.max = maxIter))
 
-                                                    return(soln)
-                                                  }, mc.cores = nCores)
-              } else {
-                solnList <- parallel::mclapply(X = hStarts,
-                                               FUN = function(h) {
-                                                 soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
-                                                                lower = 0, upper = 1e06, control = list(iter.max = maxIter))
-
-                                                 return(soln)
-                                               }, mc.cores = nCores)
-              }
+                                              return(soln)
+                                            },
+                                            n.core = n.core,
+                                            parallel.method = parallel.method,
+                                            count = verbose)
               solnNo <- which.min(unlist(lapply(solnList, function(x) x$objective)))
               soln <- solnList[[solnNo]]
             } else {
@@ -890,23 +905,17 @@ estPhylo <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.set
 
 
             if (length(hStarts) >= 2) {
-              if (verbose) {
-                solnList2 <- pbmcapply::pbmclapply(X = hStarts,
-                                                   FUN = function(h) {
-                                                     soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
-                                                                    lower = 0, upper = 1e06, control = list(iter.max = maxIter))
+              solnList2 <- parallel.compute(vec = hStarts,
+                                            func = function(h) {
+                                              soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
+                                                             lower = 0, upper = 1e06, control = list(iter.max = maxIter))
 
-                                                     return(soln)
-                                                   }, mc.cores = nCores)
-              } else {
-                solnList2 <- parallel::mclapply(X = hStarts,
-                                                FUN = function(h) {
-                                                  soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
-                                                                 lower = 0, upper = 1e06, control = list(iter.max = maxIter))
+                                              return(soln)
+                                            },
+                                            n.core = n.core,
+                                            parallel.method = parallel.method,
+                                            count = verbose)
 
-                                                  return(soln)
-                                                }, mc.cores = nCores)
-              }
               solnNo2 <- which.min(unlist(lapply(solnList2, function(x) x$objective)))
               soln2 <- solnList2[[solnNo2]]
             } else {
@@ -1330,7 +1339,28 @@ estPhylo <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.set
 #' then the `addNOIA` (or corresponding) option in the `calcGRM` function will be used,
 #' and if this argument is `diffusion`, the diffusion kernel based on Laplacian matrix will be computed from network.
 #' You can assign more than one kernelTypes for this argument; for example, kernelTypes = c("addNOIA", "diffusion").
-#' @param nCores The number of cores used for optimization.
+#' @param n.core Setting n.core > 1 will enable parallel execution on a machine with multiple cores.
+#' This argument is not valid when `parallel.method = "furrr"`.
+#' @param parallel.method Method for parallel computation. We offer three methods, "mclapply", "furrr", and "foreach".
+#'
+#' When `parallel.method = "mclapply"`, we utilize \code{\link[pbmcapply]{pbmclapply}} function in the `pbmcapply` package
+#' with `count = TRUE` and \code{\link[parallel]{mclapply}} function in the `parallel` package with `count = FALSE`.
+#'
+#' When `parallel.method = "furrr"`, we utilize \code{\link[furrr]{future_map}} function in the `furrr` package.
+#' With `count = TRUE`, we also utilize \code{\link[progressr]{progressor}} function in the `progressr` package to show the progress bar,
+#' so please install the `progressr` package from github (\url{https://github.com/HenrikBengtsson/progressr}).
+#' For `parallel.method = "furrr"`, you can perform multi-thread parallelization by
+#' sharing memories, which results in saving your memory, but quite slower compared to `parallel.method = "mclapply"`.
+#'
+#' When `parallel.method = "foreach"`, we utilize \code{\link[foreach]{foreach}} function in the `foreach` package
+#' with the utilization of \code{\link[parallel]{makeCluster}} function in `parallel` package,
+#' and \code{\link[doParallel]{registerDoParallel}} function in `doParallel` package.
+#' With `count = TRUE`, we also utilize \code{\link[utils]{setTxtProgressBar}} and
+#' \code{\link[utils]{txtProgressBar}} functions in the `utils` package to show the progress bar.
+#'
+#' We recommend that you use the option `parallel.method = "mclapply"`, but for Windows users,
+#' this parallelization method is not supported. So, if you are Windows user,
+#' we recommend that you use the option `parallel.method = "foreach"`.
 #' @param hOpt Optimized hyper parameter for constructing kernel when estimating haplotype effects.
 #'  If hOpt = "optimized", hyper parameter will be optimized in the function.
 #'  If hOpt is numeric, that value will be directly used in the function.
@@ -1409,8 +1439,8 @@ estNetwork <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.s
                        subpopInfo = NULL, groupingMethod = "kmedoids", nGrp = 3,
                        nIterClustering = 100, iterRmst = 100, networkMethod = "rmst",
                        autogamous = FALSE, probParsimony = 0.95, nMaxHaplo = 1000,
-                       kernelTypes = "addNOIA", nCores = parallel::detectCores() - 1,
-                       hOpt = "optimized", hOpt2 = "optimized", maxIter = 20,
+                       kernelTypes = "addNOIA", n.core = parallel::detectCores() - 1,
+                       parallel.method = "mclapply", hOpt = "optimized", hOpt2 = "optimized", maxIter = 20,
                        rangeHStart = 10 ^ c(-1:1), saveName = NULL, saveStyle = "png",
                        plotWhichMDS = 1:2, colConnection = c("grey40", "grey60"),
                        ltyConnection = c("solid", "dashed"), lwdConnection = c(1.5, 0.8),
@@ -1897,23 +1927,16 @@ estNetwork <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.s
 
 
             if (length(hStarts) >= 2) {
-              if (verbose) {
-                solnList <- pbmcapply::pbmclapply(X = hStarts,
-                                                  FUN = function(h) {
-                                                    soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
-                                                                   lower = 0, upper = 1e06, control = list(iter.max = maxIter))
+              solnList <- parallel.compute(vec = hStarts,
+                                            func = function(h) {
+                                              soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
+                                                             lower = 0, upper = 1e06, control = list(iter.max = maxIter))
 
-                                                    return(soln)
-                                                  }, mc.cores = nCores)
-              } else {
-                solnList <- parallel::mclapply(X = hStarts,
-                                               FUN = function(h) {
-                                                 soln <- nlminb(start = h, objective = maximizeFunc, gradient = NULL, hessian = NULL,
-                                                                lower = 0, upper = 1e06, control = list(iter.max = maxIter))
-
-                                                 return(soln)
-                                               }, mc.cores = nCores)
-              }
+                                              return(soln)
+                                            },
+                                            n.core = n.core,
+                                            parallel.method = parallel.method,
+                                            count = verbose)
               solnNo <- which.min(unlist(lapply(solnList, function(x) x$objective)))
               soln <- solnList[[solnNo]]
             } else {
@@ -2011,23 +2034,16 @@ estNetwork <- function(blockInterest = NULL, gwasRes = NULL, nTopRes = 1, gene.s
 
 
               if (length(hStarts) >= 2) {
-                if (verbose) {
-                  solnList2 <- pbmcapply::pbmclapply(X = hStarts,
-                                                     FUN = function(h) {
-                                                       soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
-                                                                      lower = 0, upper = 1e06, control = list(iter.max = maxIter))
+                solnList2 <- parallel.compute(vec = hStarts,
+                                              func = function(h) {
+                                                soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
+                                                               lower = 0, upper = 1e06, control = list(iter.max = maxIter))
 
-                                                       return(soln)
-                                                     }, mc.cores = nCores)
-                } else {
-                  solnList2 <- parallel::mclapply(X = hStarts,
-                                                  FUN = function(h) {
-                                                    soln <- nlminb(start = h, objective = maximizeFunc2, gradient = NULL, hessian = NULL,
-                                                                   lower = 0, upper = 1e06, control = list(iter.max = maxIter))
-
-                                                    return(soln)
-                                                  }, mc.cores = nCores)
-                }
+                                                return(soln)
+                                              },
+                                              n.core = n.core,
+                                              parallel.method = parallel.method,
+                                              count = verbose)
                 solnNo2 <- which.min(unlist(lapply(solnList2, function(x) x$objective)))
                 soln2 <- solnList2[[solnNo2]]
               } else {